Allow `Parser` and `BufLexer` to unwrap to their inner types
This is done by adding `into_inner(self)...` methods.

An expected use case for this is: You used `Parser` to parse some JSON
to find the position of a token within the buffer. Now in a zero-copy
manner you want get access to that token's position in the underlying
buffer. To do so: unwrap the inner `BufLexer` from the parser, and then
unwrap the buffer itself from the `BufLexer`.