Sever lifetime of lexer value from lexer itself
It was annoying that the lifetime of `Value` was tied to the lifetime of
`Lexer`. This was done entirely to support the zero-copy `BufLexer` use
case. This change eliminates the lifetime to give a cleaner more usable
interface. The trade-off is that we now have to reference count the
buffer. In simple cases of up to 39 UTF-8 bytes, as long as there is no
escape sequence in the string, we just copy it them into the `Value`
structure inline rather than copying. Bigger strings are just references
into the (now reference-counted) buffer.