Fix bug: Column count was being advanced per UTF-8 byte 🐜
It is more human-friendly to align the column index per line with input
characters, in the same way that the line index is aligned with input
lines.

This commit applies this intuition. The column count is now incremented
only after an entire valid UTF-8 byte sequence is read. This should
align the column number reported more in error messages with what a
human consuming the error message would expect.