Remove `BufAnalyzer::into_inner` to hide implementation detail
It's not really vital to have `into_inner` since users can pass either
an owned value that dereferences to a `&[u8]` or a `&[u8]` itself. If
some kind of shared ownership is desired to make a newtype that wraps
`Arc<Vec<u8>>` and implement `Deref<Target=[u8]>` for it.

So in the interests of hiding implementation details that don't need to
be public, I removed the method.