Give parser a configurable max level
The maximum level is the maximum amount of nesting the parser will allow
before it stops parsing with `ErrorKind:::Level`.

The default maximum level is currently the number of levels that will fit
in 24 bytes, which is 192.

The purpose of this change is to enable developers to fully understand
(and constrain) the parser's allocation behavior; and from a security
perspective, to allow developer's to protect their applications from
degenerate or malicious inputs.