Improve ergonomics of content fetch interface (breaking)
This commit replaces the `lexical::Analyzer::content` and
`Parser::content` methods, which used to return a `Result`, with the
following method trifecta:

1. `fn content(&self) -> Content`
2. `fn err(&self) -> Error`
3. `fn try_content(&self) -> Result<Content, Error`

The reason behind this change is that just about any parsing loop needs
to `match` on the value of `.next()` and proactively check for errors.
Since the loop code already "knows" whether it is in the success case or
the error case, only making the content/error available via `Result`
causes developers to have to use a bunch of unwieldy `.unwrap()` and
`.unwrap_err()` calls to get what they want, which adds friction to the
code and makes it harder to understand.

This change allows developers to leverage the knowledge they have about
being in error or not to avoid all those frictitious unwraps.

This is a breaking change and will require a minor version bump to
v0.3.0.