Improve `unescaped` ergonomics, shrink `buf::Content` size (breaking)
This commit contains a breaking change to the `lexical::Content` trait,
switching the return value of `lexical::Content::unescaped` from a bare
`&str` to a `Cow<'_, str>`.

The behavior of the method is also changed: if it has to unescape, it
now never caches the value. It re-computes the escaping on each call.

As a result of making this change and reducing the number of inlined
bytes from 39 to 30, the size of `buf::Content` shrinks from 48 bytes
(six 64-bit words) to 32 bytes (four 32-bit words), as reported by
`$ cargo +nightly rustc -- -Zprint-type-sizes`. This smaller size should
(hopefully) be somewhat more palatable for hot loop situations.

The goal of the main breaking change is to ensure that if callers need
an owned unescaped value, they can limit the number of allocations to
one, whereas in the previous interface there would be two: first
`Content::unescaped` would allocate and cache the unescaped value, and
return a reference to it; then the caller would have to convert that
reference into an owned value by allocating again!

Lastly, I think caching the unescaped value was a silly optimization
that increases complexity but probably has little real-world impact.
In the rare cases where a caller might really want to cache this value
to avoid re-computing the unescaping, he can do that himself.