Add more convenience methods to `Token` (breaking)
The breaking change is renaming `Token::is_value` to
`Token::is_primitive`, which is slightly more precise. The JSON RFC
describes the three literal values plus strings and numbers as being
"primitive values", and these are a subset of all values since arrays
and objects are also values.

The added convenience methods are:

- `is_literal()` (is it a literal false, null, or true?)
- `is_pseudo()` (is it a pseudo-token?)
- `is_terminal()` (is it EOF or error?)