Remove `mut` from `&self` in `Content::unescaped` (breaking)
This is lightly breaking both it can induce compiler warnings due to
unnecessary `mut` designations. I don't think it should hard break
anything, but better to treat it as an interface break.

The removal of caching from `fixed::Content` made this `&mut self`
redundant, and the interface is much cleaner without it.