Make `lexical::unescape` fully `pub`
It's now an independently-usable piece of functionality that library
users can depend on even if they don't want the whole lexical analysis
stack.

Note that I had been planning to change the return type to `Result` but
did not do so in this iteration. It might still happen in the future,
but for now `panic!` seems reasonable.